handle vscode special vars in cmake.env options
