+++ src/config.ts
@@ -1,6 +1,7 @@
 import * as os from 'os';
 import * as path from 'path';
 import * as vscode from 'vscode';
+import * as util from './util';
 
 import {Maybe} from './util';
 
@@ -9,6 +10,15 @@ export class ConfigurationReader {
     const config = vscode.workspace.getConfiguration('cmake');
     const value = config.get(key);
     return (value !== undefined) ? value as T : default_;
+  }
+
+  private _escapePaths(obj: Object) {
+    return Object.getOwnPropertyNames(obj).reduce(
+      (acc, key: string) => {
+        acc[key] = util.replaceVars(obj[key]);
+        return acc;
+      },
+      {});
   }
 
   private _readPrefixed<T>(key): T|null {
@@ -114,19 +124,19 @@ export class ConfigurationReader {
   }
 
   get environment() {
-    return this._readPrefixed<{[key: string]: string}>('environment') || {};
+    return this._escapePaths(this._readPrefixed<{[key: string]: string}>('environment') || {});
   }
 
   get configureEnvironment() {
-    return this._readPrefixed<{[key: string]: string}>('configureEnvironment') || {};
+    return this._escapePaths(this._readPrefixed<{[key: string]: string}>('configureEnvironment') || {});
   }
 
   get buildEnvironment() {
-    return this._readPrefixed<{[key: string]: string}>('buildEnvironment') || {};
+    return this._escapePaths(this._readPrefixed<{[key: string]: string}>('buildEnvironment') || {});
   }
 
   get testEnvironment() {
-    return this._readPrefixed<{[key: string]: string}>('testEnvironment') || {};
+    return this._escapePaths(this._readPrefixed<{[key: string]: string}>('testEnvironment') || {});
   }
 
   get defaultVariants(): Object {
